+++ app/src/main/java/com/example/instasnap/View/User/Adapters/StoryRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.instasnap.View.HomePage;
+package com.example.instasnap.View.User.Adapters;
 
 import android.view.LayoutInflater;
 import android.view.View;
