+++ app/src/main/java/com/example/instasnap/View/User/Fragments/SettingsFragmentView.java
@@ -1,0 +1,31 @@
+package com.example.instasnap.View.User.Fragments;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.instasnap.R;
+
+public class SettingsFragmentView extends Fragment {
+
+
+    public SettingsFragmentView() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.settings_page_fragment, container, false);
+    }
+}
