+++ app/src/main/res/layout/main_page_fragment.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".View.User.Fragments.MainPageFragmentView">
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/main_container_view"
+        android:name="com.example.instasnap.View.User.Fragments.HomePageFragmentView"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:layout="@layout/home_page_fragment" />
+</FrameLayout>
