+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,6 @@
     <string name="screen_mode">Screen Mode</string>
     <string name="settings">Settings</string>
     <string name="profile_picture">profile_picture</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
